pyproject.toml
[tool.poetry]
name = "lightech"
version = "0.1.0"
description = ""
authors = ["suband <suband74@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
drf-jwt = "^1.19.2"
gunicorn = "^22.0.0"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.8.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
